requirements.txt
Flask==2.0.1
requests==2.25.1
flask-cors==3.0.10
twilio==6.57.0
gunicorn==20.1.0
python-dotenv==0.19.2
